Refactor to merge createFromTemplate with createSwa
