+++ src/commands/createStaticWebApp/IStaticWebAppWizardContext.ts
@@ -43,5 +43,4 @@ export interface IStaticWebAppWizardContext extends IResourceGroupWizardContext,
 
     fromTemplate?: boolean;
     templateRepo?: RepoData;
-    templateRepos?: RepoData[];
 }
