+++ src/commands/createStaticWebApp/createStaticWebApp.ts
@@ -3,38 +3,25 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ProgressLocation, ProgressOptions, window } from 'vscode';
 import { IActionContext, ICreateChildImplContext } from 'vscode-azureextensionui';
 import { productionEnvironmentName } from '../../constants';
 import { ext } from '../../extensionVariables';
 import { EnvironmentTreeItem } from '../../tree/EnvironmentTreeItem';
 import { StaticWebAppTreeItem } from '../../tree/StaticWebAppTreeItem';
 import { SubscriptionTreeItem } from '../../tree/SubscriptionTreeItem';
-import { localize } from '../../utils/localize';
-import { showNoWorkspacePrompt, tryGetWorkspaceFolder } from '../../utils/workspaceUtils';
+import { verifyWorkSpace } from '../../utils/workspaceUtils';
 import { showSwaCreated } from '../showSwaCreated';
 import { IStaticWebAppWizardContext } from './IStaticWebAppWizardContext';
 import { postCreateStaticWebApp } from './postCreateStaticWebApp';
-import { setWorkspaceContexts } from './setWorkspaceContexts';
 
 export async function createStaticWebApp(context: IActionContext & Partial<ICreateChildImplContext> & Partial<IStaticWebAppWizardContext>, node?: SubscriptionTreeItem): Promise<StaticWebAppTreeItem> {
     if (!node) {
         node = await ext.tree.showTreeItemPicker<SubscriptionTreeItem>(SubscriptionTreeItem.contextValue, context);
     }
 
-    const progressOptions: ProgressOptions = {
-        location: ProgressLocation.Notification,
-        title: localize('verifyingWorkspace', 'Verifying workspace...')
-    };
-    await window.withProgress(progressOptions, async () => {
-        const folder = await tryGetWorkspaceFolder(context);
-        if (folder) {
-            await setWorkspaceContexts(context, folder);
-        } else {
-            await showNoWorkspacePrompt(context);
-        }
-    });
-
+    if (!context.fromTemplate) {
+        await verifyWorkSpace(context);
+    }
     const swaNode: StaticWebAppTreeItem = await node.createChild(context);
     void showSwaCreated(swaNode);
 
@@ -45,6 +32,10 @@ export async function createStaticWebApp(context: IActionContext & Partial<ICrea
 
     void postCreateStaticWebApp(swaNode);
     return swaNode;
+}
+
+export async function createStaticWebAppFromTemplate(context: IActionContext, node?: SubscriptionTreeItem): Promise<StaticWebAppTreeItem> {
+    return await createStaticWebApp({ ...context, fromTemplate: true }, node);
 }
 
 export async function createStaticWebAppAdvanced(context: IActionContext, node?: SubscriptionTreeItem): Promise<StaticWebAppTreeItem> {
