--- src/commands/createStaticWebApp/createStaticWebAppFromTemplate.ts
@@ -1,23 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { IActionContext, ICreateChildImplContext } from 'vscode-azureextensionui';
-import { ext } from '../../extensionVariables';
-import { StaticWebAppTreeItem } from '../../tree/StaticWebAppTreeItem';
-import { SubscriptionTreeItem } from '../../tree/SubscriptionTreeItem';
-import { showSwaCreated } from '../showSwaCreated';
-import { IStaticWebAppWizardContext } from './IStaticWebAppWizardContext';
-
-export async function createStaticWebAppFromTemplate(context: IActionContext & Partial<ICreateChildImplContext> & Partial<IStaticWebAppWizardContext>, node?: SubscriptionTreeItem): Promise<StaticWebAppTreeItem> {
-    if (!node) {
-        node = await ext.tree.showTreeItemPicker<SubscriptionTreeItem>(SubscriptionTreeItem.contextValue, context);
-    }
-
-    context.fromTemplate = true;
-    const swaNode: StaticWebAppTreeItem = await node.createChild(context);
-
-    await showSwaCreated(swaNode);
-    return swaNode;
-}
