+++ src/commands/registerCommands.ts
@@ -15,8 +15,7 @@ import { uploadAppSettings } from './appSettings/uploadAppSettings';
 import { browse } from './browse';
 import { createChildNode } from './createChildNode';
 import { createHttpFunction } from './createHttpFunction';
-import { createStaticWebApp, createStaticWebAppAdvanced } from './createStaticWebApp/createStaticWebApp';
-import { createStaticWebAppFromTemplate } from './createStaticWebApp/createStaticWebAppFromTemplate';
+import { createStaticWebApp, createStaticWebAppAdvanced, createStaticWebAppFromTemplate } from './createStaticWebApp/createStaticWebApp';
 import { createSwaConfigFile } from './createSwaConfigFile';
 import { deleteEnvironment } from './deleteEnvironment';
 import { deleteNode } from './deleteNode';
