+++ src/constants.ts
@@ -33,8 +33,6 @@ export const isEndGroup = (t: string): boolean => /##\[endgroup\]/.test(t)
 export const githubAuthProviderId: string = 'github';
 export const githubScopes: string[] = ['repo', 'workflow', 'admin:public_key'];
 
-export const templateReposUsername: string = 'staticwebdev';
-
 // Source: https://github.com/github/gitignore/blob/master/Node.gitignore
 export const defaultGitignoreContents: string = `# Logs
 logs
